Extract buildLineDiff into a shared ui/src/lib/line-diff helper
Lifts the LCS-based line diff out of DocumentDiffModal so that the routine
revision history diff modal can reuse the same algorithm without duplication.
Behavior is unchanged.
